Chrome: move the keyboard navigation to the Header Toolbar
